doc2txt/requirements.txt
asgiref==3.2.7
Django==3.0.7
docx2python==1.23
lxml==4.5.1
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
